app/requirements.txt
fastapi[standard]
sqlmodel==0.0.19
python-dotenv
psycopg2